feat: Change to a class over flask-restful

+++ app/users/views.py
@@ -4,16 +4,17 @@ from flask import request, jsonify
 from flask_restplus import Resource
 from marshmallow import ValidationError
 from .models import Users
-from .serializers import UserSerializer, UserPostSerializer
+from .serializers import UserSerializer
 from flask_login import login_user, current_user, login_required, logout_user
 from flask_jwt_extended import create_access_token
 
 
-@api.route('/get_user')
-@api.route('/get_user/<int:user_id>')
+@api.route('/')
+@api.route('/<int:user_id>')
 class User(Resource):
     method_decorators = [login_required]
-    user_serializer = UserSerializer()
+    print_serialize_fields = ['id', 'first_name', 'last_name', 'email', 'rol', 'position']
+    user_serializer = UserSerializer(only=print_serialize_fields)
 
     def get(self, user_id=None):
         if user_id:
@@ -27,37 +28,43 @@ class User(Resource):
 
     def post(self):
         if current_user.rol_id == 1:
-            self.user_serializer = UserPostSerializer()
+            self.user_serializer = UserSerializer()
             user_data = request.get_json(force=True)
             try:
                 user_serialize = Users(**self.user_serializer.load(user_data))
             except ValidationError as e:
-                return e.messages, 400
-            user_serialize.create_object()
-            return {'message': 'created'}, 201
+                return {'code': 400, 'errors':e.messages}, 400
+            if not Users.query.get(user_serialize.id):
+                user_serialize.create_object()
+                return {'code': 201, 'message': 'created'}, 201
+            else:
+                return {'code': 400, 'error': 'User with id {} already exist'.format(user_serialize.id)}, 400
         else:
-            return {'error': 'Not allowed to create new Users'}, 403
+            return {'code': 403, 'error': 'Not allowed to create new Users'}, 403
 
     def put(self, user_id):
         user_to_update = Users.query.get(user_id)
-        update_fields = request.get_json(force=True)
-        try:
-            serialize_update_fields = self.user_serializer.load(update_fields, partial=True)
-            user_to_update.update_object(serialize_update_fields)
-            return {'message': 'updated'}, 200
-        except ValidationError as e:
-            return e.messages, 400
+        if not user_to_update:
+            return {'code': 400, 'error': 'User does not exists'}, 400
+        if current_user.rol_id == 1 or current_user == user_to_update:
+            update_fields = request.get_json(force=True)
+            try:
+                serialize_update_fields = self.user_serializer.load(update_fields, partial=True)
+                user_to_update.update_object(serialize_update_fields)
+                return {'code': 200, 'message': 'updated'}, 200
+            except ValidationError as e:
+                return {'code': 400, 'errors': e.messages}, 400
+        else:
+            return {'code': 403, 'error': 'Not allowed to change this user'}, 403
 
     def delete(self, user_id):
-        user_to_delete = Users.query.filter_by(id=user_id).first()
+        user_to_delete = Users.query.get(user_id)
         if not user_to_delete:
-            return {'error': 'User does not exists'}, 400
-        user_serialize = self.user_serializer.dump(user_to_delete)
-        user_to_delete.delete_user()
-        return {'message': 'deleted', 'user': user_serialize}, 200
-
-
-
+            return {'code': 400, 'error': 'User does not exists'}, 400
+        if current_user.rol_id == 1 or current_user != user_to_delete:
+            user_serialize = self.user_serializer.dump(user_to_delete)
+            user_to_delete.delete_user()
+            return {'code': 200, 'message': 'deleted', 'user': user_serialize}, 200
 
 
 @users.route('/login', methods=['GET', 'POST'])
@@ -90,37 +97,6 @@ def login():
 def logout():
     logout_user()
     return jsonify({"message": "User logout"}), 200
-
-
-@users.route('/<int:user_id>', methods=['GET', 'POST', 'PUT', 'DELETE'])
-@users.route('/', methods=['GET', 'POST'])
-@login_required
-def get_user(user_id=None):
-    """User endpoint, actions available:
-        get a list of user, get a simple user, update a user
-        delete user.
-        :params user_id if not provided a list of all users will be returned
-                if provided actions can be applied to a specific user
-    """
-    if user_id:
-        return Users.get_delete_put_post(item_id=user_id)
-    else:
-        users = Users.get_delete_put_post(item_id=user_id)
-        if request.method == 'POST':
-            if current_user.rol_id == 1:
-                return jsonify(message='created', user=users.json), 201
-            else:
-                return jsonify(error='Not allowed to created users'), 401
-        else:
-            return jsonify(count=len(users.json), users=users.json), 200
-
-
-@users.route('/serializer', methods=['GET'])
-def user_serializer():
-    users_serializer = UserSerializer()
-    users = Users.query.all()
-    result = users_serializer.dump(users, many=True)
-    return jsonify(count=len(result), users=result), 200
 
 
 @users.errorhandler(404)
